refactor: remove deprecated IndexedDB utility functions and consolidate IndexedDBManager implementation
